agora vai

+++ nearest LCA/p2.cpp
@@ -12,18 +12,18 @@ using namespace std;
 // ------------------------------------------
 // global variables
 
-int nVertices, nEdges;
+long long nVertices, nEdges;
 bool works = true;
 
 // ------------------------------------------
 // in and out control for nodes
 struct Node {
-  int in = 0, out = 0;
+  long long in = 0, out = 0;
 } node;
 
 // ------------------------------------------
 
-void isCyclic(int vertice, vector<int> graph[], vector<bool> flag ){
+void isCyclic(long long vertice, vector<long long> graph[], vector<bool> flag ){
 
   flag[vertice] = true;
 
@@ -31,7 +31,7 @@ void isCyclic(int vertice, vector<int> graph[], vector<bool> flag ){
   // for(int x : flag) cout << x << " | ";
   // cout << endl;
 
-  for(int ancestral : graph[vertice]){
+  for(long long ancestral : graph[vertice]){
 
     if(flag[ancestral] == true) works = false;
     else if(flag[ancestral] == false){
@@ -44,7 +44,7 @@ void isCyclic(int vertice, vector<int> graph[], vector<bool> flag ){
 
 // ------------------------------------------
 
-void LCA(int yellow, int red, vector <int> graph[] , set <int> &answer){
+void LCA(long long yellow, long long red, vector <long long> graph[] , set <long long> &answer){
 
   vector <char> colorFlag(nVertices+1, 'w');
   // w -> white -> without color
@@ -55,18 +55,18 @@ void LCA(int yellow, int red, vector <int> graph[] , set <int> &answer){
   // coloring the main nodes
   colorFlag[yellow] = 'y';  
   colorFlag[red] = 'r';
-  stack <int> processor;
+  stack <long long> processor;
   vector <Node> degree(nVertices+1);
 
   // dfs starting in yellow
   processor.push(yellow);
   while (not processor.empty()) {
 
-    int top = processor.top();
+    long long top = processor.top();
     colorFlag[top] = 'y';
     processor.pop();
 
-    for(int ancestral : graph[top] ) processor.push(ancestral);
+    for(long long ancestral : graph[top] ) processor.push(ancestral);
     
   }
 
@@ -74,18 +74,18 @@ void LCA(int yellow, int red, vector <int> graph[] , set <int> &answer){
   processor.push(red);
   while (not processor.empty()) {
 
-    int top = processor.top();
+    long long top = processor.top();
 
     if(colorFlag[top] == 'y') colorFlag[top] = 'b';
     else if(colorFlag[top] == 'w') colorFlag[top] = 'r';
 
     processor.pop();
 
-    for(int ancestral : graph[top] ) processor.push(ancestral);
+    for(long long ancestral : graph[top] ) processor.push(ancestral);
   }
 
   if(debugColors){
-    for(int i=0; i<=nVertices; i++) cout << i << " ";
+    for(long long i=0; i<=nVertices; i++) cout << i << " ";
     cout << endl;
     for(char color : colorFlag) cout << color << " ";
     cout << endl;
@@ -93,7 +93,7 @@ void LCA(int yellow, int red, vector <int> graph[] , set <int> &answer){
 
   for(int i=1; i<=nVertices; i++){
     if(colorFlag[i] == 'b'){
-      for(int ancestral : graph[i]){
+      for(long long ancestral : graph[i]){
         if(colorFlag[ancestral] == 'b'){
           degree[i].in++;
           degree[ancestral].out++;
@@ -102,7 +102,7 @@ void LCA(int yellow, int red, vector <int> graph[] , set <int> &answer){
     }
   }
 
-  for(int i=1; i<=nVertices; i++){
+  for(long long i=1; i<=nVertices; i++){
     if(degree[i].out == 0 and colorFlag[i] == 'b') answer.insert(i);
   }
 
@@ -117,17 +117,17 @@ int main(){
   ios::sync_with_stdio(false);
 
   // general inputs
-  int v1, v2;
+  long long v1, v2;
   cin >> v1 >> v2;
   cin >> nVertices >> nEdges;
 
   // graph and flags alocation
-  vector <int> graph[nVertices+1];
+  vector <long long> graph[nVertices+1];
   vector <bool> flag(nVertices+1, 0);
 
   // graph input
-  for(int i=0; i<nEdges; i++){
-    int dad, son;
+  for(long long i=0; i<nEdges; i++){
+    long long dad, son;
     cin >> dad >> son;
 
     graph[son].push_back(dad);
@@ -138,7 +138,7 @@ int main(){
   // cout << endl;
 
   // genealogic exeptions treatment
-  for(int i=nVertices; i>0; --i){
+  for(long long i=nVertices; i>0; --i){
     isCyclic(i, graph, flag);
     if ( (graph[i].size() > 2) or not works) {
       cout << "0" << endl;
@@ -147,12 +147,12 @@ int main(){
   }
 
   // FINALLY THE NEAREST LCA
-  set <int> nearestAncestrals;
+  set <long long> nearestAncestrals;
   LCA(v1, v2, graph, nearestAncestrals);
 
 
   // found any LCA?
-  for(int valid : nearestAncestrals){
+  for(long long valid : nearestAncestrals){
     cout << valid << " ";
   }
